de components a pages
