--- src/front/components/NotFound.jsx
@@ -1,15 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import { VistaHome } from "./VistaHome";
-
-export const NotFound = () => {
-
-  const navigate = useNavigate();
-
-  const rutaVistaHome =() => {
-    navigate("/vistahome");
-  };
-
-  return (
-      <button className="text-black bg-yellow-700" onClick={rutaVistaHome}>Home</button>
-  );
-};
