+++ src/front/pages/Evento.jsx
@@ -1,0 +1,15 @@
+import { useNavigate } from "react-router-dom";
+import { VistaHome } from "../components/VistaHome";
+
+export const Evento = () => {
+
+  const navigate = useNavigate();
+
+  const rutaVistaHome = () => {
+    navigate("/vistahome");
+  };
+
+  return (
+    <button className="text-black bg-yellow-700" onClick={rutaVistaHome}>Home</button>
+  );
+};
