+++ src/front/routes.jsx
@@ -1,9 +1,9 @@
 // Import necessary components and functions from react-router-dom.
 
 import {
-    createBrowserRouter,
-    createRoutesFromElements,
-    Route,
+  createBrowserRouter,
+  createRoutesFromElements,
+  Route,
 } from "react-router-dom";
 import { Layout } from "./pages/Layout";
 import { Home } from "./pages/Home";
@@ -11,36 +11,36 @@ import { Single } from "./pages/Single";
 import { Demo } from "./pages/Demo";
 import { Login } from "./components/Login";
 import { VistaHome } from "./components/VistaHome";
-import { Formulario } from "./components/Formulario";
-import { Evento } from "./components/Evento";
-import { Usuario } from "./components/Usuario";
-import { AjustesUsuario } from "./components/AjustesUsuario";
-import { NotFound } from "./components/NotFound";
+import { Formulario } from "./pages/Formulario";
+import { Evento } from "./pages/Evento";
+import { Usuario } from "./pages/Usuario";
+import { AjustesUsuario } from "./pages/AjustesUsuario";
+import { NotFound } from "./pages/NotFound";
 
 export const router = createBrowserRouter(
-    createRoutesFromElements(
+  createRoutesFromElements(
     // CreateRoutesFromElements function allows you to build route elements declaratively.
     // Create your routes here, if you want to keep the Navbar and Footer in all views, add your new routes inside the containing Route.
     // Root, on the contrary, create a sister Route, if you have doubts, try it!
     // Note: keep in mind that errorElement will be the default page when you don't get a route, customize that page to make your project more attractive.
     // Note: The child paths of the Layout element replace the Outlet component with the elements contained in the "element" attribute of these child paths.
 
-      // Root Route: All navigation will start from here.
-      <Route path="/" element={<Layout />} errorElement={<h1>Not found!</h1>} >
+    // Root Route: All navigation will start from here.
+    <Route path="/" element={<Layout />} errorElement={<h1>Not found!</h1>} >
 
-        {/* Nested Routes: Defines sub-routes within the BaseHome component. */}
-        <Route path= "/" element={<Home />} />
-        <Route path="/single/:theId" element={ <Single />} />  {/* Dynamic route for single items */}
-        <Route path="/demo" element={<Demo />} />
-        <Route path="/home" element={<home />} />
-        <Route path="login" element={<Login />} />
-        <Route path="vistahome" element={<VistaHome />} />
-        <Route path="formulario" element={<Formulario />} />
-        <Route path="evento" element={<Evento />} />
-        <Route path="usuario" element={<Usuario />} />
-        <Route path="ajustesusuario" element={<AjustesUsuario />} />
-        <Route path="notfound" element={<NotFound />} />
+      {/* Nested Routes: Defines sub-routes within the BaseHome component. */}
+      <Route path="/" element={<Home />} />
+      <Route path="/single/:theId" element={<Single />} />  {/* Dynamic route for single items */}
+      <Route path="/demo" element={<Demo />} />
+      <Route path="/home" element={<home />} />
+      <Route path="login" element={<Login />} />
+      <Route path="vistahome" element={<VistaHome />} />
+      <Route path="formulario" element={<Formulario />} />
+      <Route path="evento" element={<Evento />} />
+      <Route path="usuario" element={<Usuario />} />
+      <Route path="ajustesusuario" element={<AjustesUsuario />} />
+      <Route path="notfound" element={<NotFound />} />
 
-      </Route>
-    )
+    </Route>
+  )
 );
